Move GetPath_FromExecutingAssembly_ShouldMatchLocation to test separate file
