+++ test/Tethos.Tests/AssemblyExtensionsTests.cs
@@ -193,20 +193,6 @@ namespace Tethos.Tests
             actual.Should().HaveCount(expected);
         }
 
-        [Fact]
-        public void GetPath_FromExecutingAssembly_ShouldMatchLocation()
-        {
-            // Arrange
-            var assembly = Assembly.GetExecutingAssembly();
-            var expected = assembly.Location;
-
-            // Act
-            var actual = expected.GetFile();
-
-            // Assert
-            actual.Path.Should().Be(expected);
-        }
-
         [Theory, AutoData]
         internal void ElseLoadReferencedAssemblies_ShouldReturnOriginal(File[] files)
         {
